Xcode 5 project update
It will probably break the build for versions prior to 5, sorry guys,
but since mavericks is free now, I believe everyone should update it

+++ toolkit/Source/src/editor/EditorBase.macosx.mm
@@ -59,7 +59,7 @@ static bool OpenPanel(
 	NSURL* initPath = [NSURL URLWithString:[NSString stringWithUTF8String:initDirectory.c_str()]];
 	[panel setDirectoryURL:initPath];
 
-	const int i = [panel runModal];
+	const long i = [panel runModal];
 	const bool r = (i == NSFileHandlingPanelOKButton);
 
 	if (r)
