+++ toolkit/Source/src/engine/Platform/FileListing.h
@@ -48,7 +48,7 @@ public:
 
 	inline unsigned int GetNumFiles()
 	{
-		return m_fileName.size();
+		return static_cast<unsigned int>(m_fileName.size());
 	}
 
 	inline bool GetFileName(const unsigned int index, FILE_NAME& fileName)
